requirements.txt
blinker==1.3
cffi==1.11.4
chardet==2.3.0
click==6.7
configobj==5.0.6
cryptography==1.2.3
dj-database-url==0.4.2
Django==2.0.1
djangorestframework==3.7.7
drf-nested-routers==0.90.0
gunicorn==19.7.1
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
MarkupSafe==0.23
oauthlib==1.0.3
prettytable==0.7.2
psycopg2==2.7.3.2
pyasn1==0.1.9
pycparser==2.18
PyJWT==1.3.0
pyserial==3.0.1
python-dotenv==0.7.1
pytz==2017.3
PyYAML==3.11
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
urllib3==1.13.1
virtualenv==15.1.0
